Reworked core functions. Increased execution speed.

+++ apis/requests/CancelOrder.go
@@ -1,5 +1,5 @@
 package requests
 
 type CancelOrderSettings struct {
-	OrderID int `json:"order_id"` // order ID
+	OrderID uint64 `json:"order_id"` // order ID
 }
