+++ apis/requests/OrderInfo.go
@@ -1,5 +1,5 @@
 package requests
 
 type OrderInfoSettings struct {
-	OrderID int `json:"order_id"` // order ID
+	OrderID uint64 `json:"order_id"` // order ID
 }
