+++ responses/Depth.go
@@ -1,14 +1,14 @@
 package responses
 
 type Depth struct {
-	Success  int `json:"success"`
+	Success  uint8 `json:"success"`
 	PairData map[string]PData
 	Error    string `json:"error"`
 }
 
 type PData struct {
-	Asks [][2]float64 `json:"asks"` // ордера на продажу
-	Bids [][2]float64 `json:"bids"` // ордера на покупку
+	Asks [][2]float64 `json:"asks"` // selling orders
+	Bids [][2]float64 `json:"bids"` // buying orders
 }
 
 func NewDepth() Depth {
