+++ responses/Info.go
@@ -1,7 +1,7 @@
 package responses
 
 type Info struct {
-	Success    int                               `json:"success"`
+	Success    uint8                             `json:"success"`
 	ServerTime uint64                            `json:"server_time"`
 	Pairs      map[string]map[string]interface{} `json:"pairs"`
 	Error      string                            `json:"error"`
